--- src/test/java/com/alexlis/pages/confirmActionsPages/ConfirmActionsPage.java
@@ -1,26 +1,0 @@
-package com.alexlis.pages.confirmActionsPages;
-
-import com.alexlis.pages.PageObjects;
-import com.codeborne.selenide.SelenideElement;
-
-import static com.codeborne.selenide.Selectors.byText;
-import static com.codeborne.selenide.Selenide.$;
-import static io.qameta.allure.Allure.step;
-
-public class ConfirmActionsPage{
-
-//    private SelenideElement
-//            ageButtonYes = $("[data-action=confirm]").as("Кнопка ДА"),
-//            regionButtonYes = $(byText("Да, верно"));
-//
-//    public PageObjects confirmAge() {
-//        step("Подтверждение региона", () ->
-//                ageButtonYes.click());
-//        return this;
-//    }
-//
-//    public void regionConfirm() {
-//        step("Подтверждение региона", () ->
-//                regionButtonYes.click());
-//    }
-}
