--- src/test/java/com/alexlis/pages/elementsPages/NavigationElementsPages.java
@@ -1,23 +1,0 @@
-package com.alexlis.pages.elementsPages;
-
-
-import com.alexlis.pages.PageObjects;
-import com.codeborne.selenide.SelenideElement;
-
-import static com.codeborne.selenide.Selenide.open;
-import static io.qameta.allure.Allure.step;
-
-public class NavigationElementsPages {
-
-    private final String URL = "https://www.34play.me/";
-
-private SelenideElement  a;
-
-//
-//    public PageObjects openPage() {
-//        step("Открыть страницу", () ->
-//                open(URL));
-//        return this;
-//    }
-
-}
